Revert Anbernic bricking uboot on SD card when using F button to boot into Android
For those with issues booting back into Linux when using GammaOS and using the F button to boot into Android, install this Magisk module in Android, then reboot twice. 
You'll be back in Linux. Going forward with this Magisk module installed, you won't need to worry about whether you hold the F button or not, it will work properly. 

This will be baked into a future GammaOS update by default, but you can use this for now. Blame Anbernic for this mess.